% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply Two Numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{Another number.}
}
\value{
The product of x and y.
}
\description{
This function takes two numbers and returns their product.
}
\examples{
multiply(4, 5)
}
